RED - Create GameOfLife module
In TDD, there's the notion of "Changing the message." The failure
message from the previous test was:

    1) Error:
    TestTransitionAtEndOfDay#test_a_con_standing_alone_dies_at_end_of_day:
    NameError: uninitialized constant TestTransitionAtEndOfDay::GameOfLife
        test/all.rb:7:in `test_a_con_standing_alone_dies_at_end_of_day'

This tells me my next step is to define a constant named `GameOfLife`.
To change the message, I defined a module `GameOfLife` in the same file.
"The same file?!" you say! "Why?!"

Because the goal of TDD is to get from Red to Green as quickly as
possible, and do the cleanup in the refactor pass.

I went with a module instead of a class mostly out of habit, as I tend to keep my
methods "stateless" for as long as possible.

A stateless method takes a set of arguments and returns a value, but
doesn't reference other data, such as instance or global variables.