GREEN - transition_at_end_of_day returns empty array
The failing message was:

    1) Failure:
    TestTransitionAtEndOfDay#test_a_con_standing_alone_dies_at_end_of_day
    [test/all.rb:15]:
    Expected: []
      Actual: nil

Now we're getting into a message that is about a difference in our
expected behavior, not the structure of our code. We expected the method
under test to return an empty set given the inputs we provided; however
it's returning nil.

So let's make it return an empty array.

"But this doesn't have any logic! It'll totally fail under pretty much
any other use case!"

To which I reply: Yes! Yes it will! That's the point. To force the code
to work well we will need to write further tests against more use cases.

When we do the least we possibly can to make the test pass it encourages
us to write tests which cover more use cases. This results in a better
tested application! Woo!