REFACTOR - Extract GameOfLife to lib directory
Refactoring is to change the structure of code without changing it's
behavior.

It's generally not a good idea to store your implementation code in the
same file as your test code. This commit moves the implementation code
into the `lib` directory.

To make this work, I have to tell Ruby to look in my projects `lib`
directory for source code. I do this by `unshift`ing (or prepending) the
absolute path to `lib` into the `$LOAD_PATH` array.

`$LOAD_PATH` is a list of directories on your computers file system
where Ruby looks for files to `require`.

When we call `require "game_of_life"` Ruby now looks in `lib` for a file
named `game_of_life.rb` and loads it as ruby code. This allows us to put
the `GameOfLife` module we defined in `lib/game_of_life.rb` and
reference it from our test code.