REFACTOR - Extract test helper for next_day and target_con
This may be a bit premature, but I noticed the possibility for a pattern
in the tests.

By making the test classes name describe a specific cell, I can define
`target_cell` and use that in both the current tests, which makes the
assert and refute statements.

Again, this is probably premature; but when you want to write a ton of
tests for very specific things, creating your own little helpers is
often quite powerful.