REFACTOR - Custom assertions for a con in a population
There was starting to be a touch of duplication in the assert/refute
stuff. This duplication led me to see a new idea that we could express
in our code: The notion of asserting a cell exists within a population.

Our test code is still our code, and if it lacks a design that allows
for change, can frequently get in the way of producing useful software.