REFACTOR - extract count_neighbors method
By moving this logic into another function we're expressing yet
another importan idea.

Sure the function is a bit of a hack as it doesn't take into account
*which cons* neighbors to worry about. We'll get to that.

Now we have a choice in our next test:

1. Continue driving features through the outside edge of `next_day`
2. Begin testing the count_neighbors sub-component

Some prefer to keep testing at the outside edge until it gets awkward or
painful.

Some prefer to start test driving out all the behavior of a
sub-component once it becomes identified; that way we can rest assured
that component works instead of having to worry about having enough high
level test cases.

It's up to the implementer... Hmmm...