RED - Sketch testing an empty world after a single iteration
Expected Output
===============
>    1) Error:
>    TestGameOfLife#test_an_empty_world_after_a_single_iteration:
>    NameError: uninitialized constant TestGameOfLife::GameOfLife
>        test/all.rb:5:in `test_an_empty_world_after_a_single_iteration'
===============

A "Sketch" test is when I write a test that with an outline of the code
I think I want. I don't really know if I'll stick with the general
shape, but it's a start that will force me to write some code.

I think this test is going to make me to write code that:

1. Defines a GameOfLife class
  * With an initialize method that takes two arguments
2. Defines an Iterator class (maybe?)
  * With an initialize method that takes 1 argument
3. Defines a `run!` instance method on `GameOfLife`
4. Defines a `world` instance method on GameOfLife
  * That returns an empty array

To be honest, this is quite a large test! I probably should have taken a
smaller step. Possibly by removing the need for an Iterator class.

Arrange/Act/Assert
==================
This test follows the "Arrange/Act/Assert" pattern:

1. We "arrange" or set-up our code by creating a GameOfLife with an
   empty world and an Iterator with the number 1.
2. We "act" by calling `run!` on the GameOfLife we created.
3. We "assert" the world in GameOfLife is stil empty.

Syntax Breakdown
================

`Iterator.new(1)` - Tell the `Iterator` class to build an "instance"
from it's blueprint, and to provide the number 1 to the instance after
it's built.

`[]` - Create an empty array.

`GameOfLife.new([], Iterator.new(1))` - Tell the `GameOfLife` class to
build an instance from it's blueprint, and to provide an empty array and
an instance of Iterator.

`game_of_life = ...` - Create a variable named "game_of_life" and store
the result of the expression on the right in it.

`game_of_life.run!()` - Call `run!` on `game_of_life` with no
arguments". Tell the data stored in the variable named
`game_of_life` to execute a method named `run!` without any additional
data.

`game_of_life.world - Call `world` on `game_of_life` with no arguments.
Note the lack of parenthesis? Ruby *implictely* calls methods if you
don't provide parenthesis. I tend to include them, but here I forgot.
Sorry!

`assert_equal(game_of_life.world, [])` - Fail the test if the result of
the exprssion `game_of_life.world` is not an empty array.

Glossary
========
"class" - A "class" is kind of like a blueprint. Something created from
the blueprint will behave in the way the blueprint says it will behave
(unless modifications are made after it's created)

"instance" - An "instance" is kind of like the result of a blueprint. It
may call all the methods and variables that are defined in the class.

"object" - A generic term for a piece of data that also has methods
available on it.

"array" - a data structure that is like a list. It may contain many
other pieces of data.

"variable" - A named location to store data for use in our program.

"expression" - A bit of code that executes and returns a value.

"value" - The actual bit of data.

"data" - A made up concept to describe things that don't actually exist,
like numbers or words or letters or formulas.