RED - Define Iterator class
Expected Output of `ruby test/all.rb`
=====================================
>    TestGameOfLife#test_an_empty_world_after_a_single_iteration:
>    ArgumentError: wrong number of arguments (1 for 0)
>        test/all.rb:13:in `initialize'
>        test/all.rb:13:in `new'
>        test/all.rb:13:in `test_an_empty_world_after_a_single_iteration'

Did you guess I would define an `Iterator` class after the last commit? Yep!
Why?  Because the error was there is nothing named Iterator, so the easiest,
fastest thing is to simply create something named Iterator!

But why a class? I'm defining `Iterator` as a class because I know the ruby
idiom that `Something.new` means the data stored in `Something` is likely a
class.

Glossary
========

`idiom` - A convention followed in a particular group or community. Each
programming language has it's own idioms, and it's best to learn them so your
code is easier to read and understand. Sometimes it makes sense to *break* an
idiom, for instance I prefer to use `()` every time I call a function, while
the ruby idiom is to *not* use `()` unless necessary