RED - Define an initializer for the Iterator
Expected Result of `ruby test/all.rb`
=====================================
>      1) Error:
>    TestGameOfLife#test_an_empty_world_after_a_single_iteration:
>    ArgumentError: wrong number of arguments (2 for 0)
>        test/all.rb:14:in `initialize'
>        test/all.rb:14:in `new'
>        test/all.rb:14:in `test_an_empty_world_after_a_single_iteration'

Wait, why did I do that?! Let's look at the previous error message:
>    TestGameOfLife#test_an_empty_world_after_a_single_iteration:
>    ArgumentError: wrong number of arguments (1 for 0)
>        test/all.rb:13:in `initialize'
>        test/all.rb:13:in `new'
>        test/all.rb:13:in `test_an_empty_world_after_a_single_iteration'

There were three things I noticed from this test failure:

1. The phrase `test/all.rb:13` - This is the location in *my code* that is
   causing the failure.
2. `Wrong number of arguments (1 for 0)` - This tells me to look for a method
   call that takes a single argument (in this case `Iterator.new(1)`). The fact
   that it's taking the number 1 as it's one argument is incidental.
3. `in initialize` - I know in ruby that `initialize` is a method I define to
   "set up" an instance of a class.

These indicators caused me to define an instance method named `initialize`
which takes one argument. I named that argument `quantity` because my intent is
for that number to be used to control how many times the iterator runs. To be
honest, this is fluff. I totally should have left the Iterator out! It's an
unnecessary component at this point and time! Ah well. Lesson learned (maybe).

An `initialize` method is where we "set up" instances of a class. A former
student used the phrase "initialize is where an instance gets it's s***
together." An accurate if somewhat uncouth explanation.

You'll notice we're not doing anything in the body of the initialize method.
This is because we're doing the least we can to make the test pass! It may be
tempting to go implement all the things, but it will derail us from getting to
a passing test.

Syntax Breakdown
================
`def initialize(quantity)` - Define an instance method called
`initialize` that takes a single argument, named `quantity` on the class
`Iterator`. `initialize` is a method that gets called as part of building an
instance of a class.