GREEN - Test passes by returning `[]` from world
Expected Result of `ruby test/all.rb`
=====================================
>  Finished in 0.001329s, 752.4454 runs/s, 752.4454 assertions/s
>  .
>
>  1 runs, 1 assertions, 0 failures, 0 errors, 0 skips

Previous Result of `ruby test/all.rb`
=====================================
>  1) Failure:
>  TestGameOfLife#test_an_empty_world_after_a_single_iteration [test/all.rb:23]:
>  Expected: nil
>    Actual: []

How did I know to do this?

1. `[test/all.rb:23]` tells me which line of my test is failing.
2. On that line is `assert_equal(game_of_life.world, [])`
3. This tells me that the expected result of calling `world` on `game_of_life`
   is an empty array
4. But unfortunately it's returning `nil` (in ruby, the result of the last
   expression is returned. Because the body of the `world` method is empty, it
   returns `nil`

By adding `[]` to my method body, `GameOfLife#world` will now always return an empty array.

Syntax Breakdown
================
`GameOfLife#World` - This is shorthand for "The instance method world on the
class GameOfLife." You'll see this frequently in ruby documentation.

`nil` - Means "nothing". `nil` is similar to 0 in that it is a thing which has
meaning, and that meaning is nothing.