REFACTOR - Rename Iterator to TurnTaker
Expected Result of `ruby test/all.rb`
=====================================
>    # Running:
>
>    .
>
>    Finished in 0.001032s, 968.9922 runs/s, 968.9922 assertions/s.
>
>    1 runs, 1 assertions, 0 failures, 0 errors, 0 skips

I decided that Iterator doesn't really reflect my intent of the Iterator
class, so I renamed it to "TurnTaker"! Yea, it's a bit awkward, but hey,
this is Ruby. A kingdom of nouns!

Notice how the tests still pass despite changing the name of the class?
This is known as "refactoring." In order for something to truly be a
refactor it must:

1. Not add any new functionality (sometimes this is harder than it
   sounds!)
2. Not break the tests.

"Rename class" is a very common refactoring. To do it you must:

1. Find all usages of the class name in your code (I use `grep` or my
   editors "find all" feature.
2. Replace every usage of the old class name with the new class name
3. Watch your tests pass! If they fail, then you likely missed a class
   name.

If you use an IDE like RubyMine, it has automated tools to do this
refactoring for you.