RED - Test that a con at 0,0 with two neighbors survives
Expected Result of running `ruby test/all.rb`
=============================================
>    1) Failure:
>  TestGameOfLife#test_a_world_with_a_con_at_0_0_with_two_neighbors_keeps_the_con_with_the_neighbors_after_a_single_iteration [test/all.rb:32]:
>  The Con at 0,0 didn't survive!

Whew, that's a really long test name! Why the heck did you do that?

1. Tests which are *specific, individual use cases* are critical when test
   driving. It lets you make a very specific bit of code work without worrying
   about the bits of functionality that come later.
2. The test name is reported with the failure, which can be useful for debugging.
3. Letters are free! You don't have to buy vowels!

You'll notice that this test differs from the previous test in both the Arrange
and the Assert sections. This test has a starting world which is a collection of
Cons, whereas the previous test had an empty starting world.

The Assert section has changed from an `assert_equals` to `assert`. I did this
because I *only* want to check if the world contains the one Con I care about.
While it is tempting to write a test that compares the worlds in their
entirety, it may result in tests which fail for silly reasons.

I've decided on an x/y coordinate plane because I find that easier to reason
about than columns and rows.

Syntax Breakdown
================
`{ :x => 0, :y => 0 }` (and similar) - Here we're creating a Hash, which is a
kind of list that uses names to store and retrieve values. Try it out:
```
con = { :x => 5, :y => 1 }
con[:x] # Retreive the value stored at :x. This evaluates to 5
con[:y] # Retrieve the value stored at :y. This evaluates to 1
con[:z] # Retrieve the value stored at :z,.This evaluates to nil (Because there is no value stored at :z)
con # Retrieve the entier hash. This evaluates to { :x => 5, :y => 1 }
con.keys() # This evaluates to [:x, :y]
con.values() # This evaluates to [5, 1]
```

`game_of_life.world.include?(...)` - Because the result of `world` is an array,
we can call "Array#include?" on it to check if it contains a value.
 * Docs: http://www.ruby-doc.org/core-2.1.5/Array.html#method-i-include-3F

`assert(...,...)` - The first argument to assert is an expression that
evaluates to a true or false value. The second is a message to display when the
test fails.