GREEN - A cell at 0,0 with two neighbors survives
This is what's known as "fakin' it til we make it." Since a goal of
TDD is to stay red for as short a time as possible, it's totally OK
to "cheat" to get from red to green.

This allows our implementation/design to emerge as we add use cases, as
well as gives us more frequent opportunities to refactor.