RED - A con at 0,0 with no neighbors dies
Expected Result of running `ruby test/all.rb`
=============================================
>  1) Failure:
>  TestGameOfLife#test_a_con_at_0_0_with_no_neighbors_dies_after_a_single_iteration
>  [test/all.rb:46]:
>  The Con at 0,0 survived!
>
>  3 runs, 3 assertions, 1 failures, 0 errors, 0 skips

We're settling into a pattern of defining a situation of what will
happen to this poor little con at 0, 0. I chose to write this test for a
few reasons:

1. I thought it would fail (And it does!)
2. Using the same con with different neighbor counts lets us avoid
   worrying about writing code to handle different locations in the
   world (for now)

Syntax Breakdown
================

`refute` - The opposite of assert. Causes tests to fail if the first
argument is true.