GREEN - Con at 0,0 with no neighbors dies.
Expected Result of running `ruby test/all.rb`
=============================================
>
>  # Running:
>
>  ...
>
>  Finished in 0.001306s, 2297.0904 runs/s, 2297.0904 assertions/s.
>
>  3 runs, 3 assertions, 0 failures, 0 errors, 0 skips

Again, our goal is to quickly move from red to green. Here we're
checking against the number of cons in the world, and *if* there are
exactly three we return the original world. If there aren't, we return a
cold and lifeless world; devoid of cons.

Syntax Breakdown
================

`<condition> ? <true value> : <false value>` - This is known as a
"ternary operation", and is short-hand for an if/else block. One should
only use ternaries to return a specific value not for larger control
flow operations. Ternaries quickly become difficult to follow when they
get more complex.