REFACTOR - Extract constant for number of neighbors
Extracting a constant is a very common way to provide meaning to a
number or string. While we may (for now) know what `3` means because we've
explored the problem domain, we may forget or have to bring on someone
less familiar. Giving a name to `3` let's us  not worry about
forgetting.

I call this "Tom Bombadilling." Giving something a name so we have power
over it.