REFACTOR - Initialize all the cons when the world is created
Initializing objects can be a "heavy" operation, so we're doing it at the
beginning.

In our GameOfLife initialize we're using ruby's `map` method to convert an array of Hashes to an
Array of Cons. Fancy, eh?

In our `world` method, we're using `map` to convert an array of Cons to
an array of Hashes. This prevents us from needing to change any of our
tests.

We want to keep our tests focused *solely on the behavior*, not on the
structure of the underlying code. You know this is true when you can
introduce classes or rename methods without your tests breaking.

Syntax Breakdown
================

`&:location` - Tells an enumerable method to call a method named
:location on item(s) instead of a passed in `do ... end` block.