REFACTOR - Move `survives?` onto Con.
Individual cons can now determine whether or not they survive. Part of
me is unhappy that `Con`s are responsible for deciding whether they live
or die, as it seems like something that should be outside of their
responsibility. Another part of me is happy that it's no longer the
`GameOfLife` class making these decisions.

We'll see where it goes from here.