REFACTOR - Extract `near?` method
I didn't like that complex chunk of boolean logic and magic numbers, so
I extracted a couple functions.

In doing so, I think I spotted a bug! (Can you guess what it is?)

Yep! You can be near on :x *OR* near on :y and you count as a neighbor!
Guess it's time to write another test!