Red - Defines an initialize method on GameOfLife
Now we're getting somewhere! We've defined the initialize method and
stated it takes a single argument named population.

Arguments are how we provide data to a method. The names used in an
argument and the names used to call the argument do not need to match,
but in this case they happen to.

Let's see what the tests think about our changes:

Expected Result of `bundle exec ruby test/all.rb`
=================================================
> Run options: --seed 36241
>
> # Running:
>
> E
>
> Finished in 0.001210s, 826.4463 runs/s, 0.0000 assertions/s.
>
>   1) Error:
> TestGameOfLife#test_a_con_at_5_5_with_no_neighbors_dies_after_one_day:
> NoMethodError: undefined method `run' for #<GameOfLife:0x007fea59aa8e80>
>     test/all.rb:14:in `test_a_con_at_5_5_with_no_neighbors_dies_after_one_day'
>
> 1 runs, 0 assertions, 0 failures, 1 errors, 0 skips

Now we get a NoMethodError, what does this one tell us?

> test/all.rb:14:in `test_a_con_at_5_5_with_no_neighbors_dies_after_one_day'
Tells us the code causing the error is at line 14 in test/all.rb

> NoMethodError: undefined method `run' for #<GameOfLife:0x007fea59aa8e80>
Tells us there is not a method named "run" defined on the GameOfLife class.

When we look at lines 13 and 14 in test/all.rb we see:

13 game = GameOfLife.new(population)
14 game.run(1)

The variable named game stores an instance of GameOfLife.new and that seems to
be working, yet on 14 we call run *on* the variable named game *with* the number 1.

Because our GameOfLife class doesn't have an instance method named run, this is
causing our test to fail.

To fix this we have to define a run method on the GameOfLife class. Seeing the
run method is expected to take one argument, we *could* include the argument in
our initial definition, or we could define the method and let the tests tell us
to define the argument. Different practitioners of TDD have different opinions
about this. Personally, it doesn't make much of a difference. Do what feels
right. If something bad happens take a smaller step.