RED - Defines population instance method on GameOfLife
Surely, surely we must have done everything we need to get this test green,
right?

Nope! We're still at a failing test! Good gravy! How! What! As we program, we
deceive ourselves into thinking we know how much work it takes to make working
software. Even taking the most direct route from Red to Green can be several
steps, *even without implementing any of the underlying behavior*! This is why
it's so important to aggressively avoid doing more work than necessary when
switching from Red to Green. The longer it takes to transition between those
states the more likely we'll introduce some amount of confounding or confusing
logic in our implementation.

So, what are the tests telling us?

Expected Result of `bundle exec ruby test/all.rb`
=================================================
> Run options: --seed 663
>
> # Running:
>
> E
>
> Finished in 0.001065s, 938.9671 runs/s, 0.0000 assertions/s.
>
>   1) Error:
> TestGameOfLife#test_a_con_at_5_5_with_no_neighbors_dies_after_one_day:
> NoMethodError: undefined method `include?' for nil:NilClass
>     test/all.rb:25:in `test_a_con_at_5_5_with_no_neighbors_dies_after_one_day'
>
> 1 runs, 0 assertions, 0 failures, 1 errors, 0 skips

Looks like there's an undefined method include? for nil:NilClass. What should we do?

If we were to follow our previous pattern of defining a method on the indicated
class we would define a method on NilClass. How would that even work?

Instead, we need to look at the line of code our test points out, line 25 in
test/all.rb:

> refute(game.population.include?(con), failure_message)

Here we see the word "include?" and the the parenthesis indicate it's a method.
We also notice there is a `.` between include? and population, which tells us
include? is being called *on* population (and with `con`).

But what is population? It doesn't have any parenthesis, so is it not a method?
It has a `.`, which indicates it *is* *on* game...

In fact, it is a method. In Ruby, it's common for methods that don't take any
arguments to not include their parenthesis both when being called and when
being defined.

To break down how this line is interpreted we'd see that:

1. The method population is called on game.
2. The method include? is called on the result of the population method with a con.
3. The result of calling game.poulation.include?(con) is passed as the first argument to refute.
4. refute checks if the result was true and if it is, fails the test with the
   message given as it's second argument.

As we become better programmers in a given language, we want to constantly
improve the accuracy and speed with which we can understand a line of code such
as this, without succumbing to over-precision. It can be difficult to choose
where to draw the line of precision, especially when our understanding of a
line of code is inaccurate.

Given the above breakdown and the error message we may determine this failure
message is caused by the population method returning a value of nil.

Given a method named include? is called on the result of population, we may
ascertain that `population` could return an Array (or something else that
responds to include?, but we'll stick with Array for now).

Given the refute statement, we may infer the Array returned *cannot include*
the con we started with.

Let's make population return an empty Array.