GREEN - Alone Con at 5,5 dies after 1 day
WOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO

Expected Result of `bundle exec ruby test/all.rb`
=================================================
> Run options: --seed 62774
>
> # Running:
>
> .
>
> Finished in 0.001340s, 746.2687 runs/s, 746.2687 assertions/s.
>
> 1 runs, 1 assertions, 0 failures, 0 errors, 0 skips

One of the idioms I dislike in Ruby is implicit returns. This means that
the *last expression evaluated in a method is returned from that
method*. I much prefer typing the word `return`, but it's important to
conform to the communities preferred style when writing code.